mengganti nama session jadi masuk

+++ hapus.php
@@ -7,14 +7,22 @@ if (!isset($_SESSION["login"])) {
     exit;
 }
 
-$conn = mysqli_connect("localhost", "root", "", "dbicam");
+$conn = mysqli_connect("localhost", "root", "", "dbicam");;
 
-$id = $_GET["id"];
+if (isset($_GET["id"])) {
 
-mysqli_query($conn, "DELETE FROM daftar_nama WHERE id = '$id'");
+    $id = $_GET["id"];
+    mysqli_query($conn, "DELETE FROM daftar_nama WHERE id = '$id'");
 
-echo "<script>
-alert('Data berhasil di hapus');
-document.location.href = 'index.php';
-</script>";
-exit;
+    echo "<script>
+    alert('Data berhasil di hapus');
+    document.location.href = 'index.php';
+    </script>";
+    exit;
+} else {
+    echo "<script>
+    alert('Data gagal di hapus');
+    document.location.href = 'index.php';
+    </script>";
+    exit;
+}
